+++ Frontend/src/App.jsx
@@ -1,13 +1,13 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Trips from "./pages/Trips";
+import Trips from "./Components/TipsDetails/Trips";
 import MainNavigation from "./Components/Navigation/MainNavigation";
-import About from "./pages/About";
+import About from "./pages/Home/About";
 // import GoogleMap from "./Components/map/GoogleMap";
-import TripDetails from "./pages/TripDetails";
 import { TripFormProvider } from "./Contexts/TripFormContext";
 import Footer from "./Components/Footer/Footer";
-import Auth from "./pages/Auth";
-import Blog from "./pages/Blog";
+import Auth from "./pages/Auth/Auth";
+import Blog from "./pages/Home/Blog";
+import TripDetails from "./Components/TipsDetails/TripDetails";
 
 function App() {
   return (
@@ -16,20 +16,18 @@ function App() {
         <Router>
           <MainNavigation />
           <div className="main-content">
-          <Routes>
+            <Routes>
               <Route path="/" exact element={<Trips />} />
               <Route path="/about" exact element={<About />} />
               <Route path="/tripDetails" exact element={<TripDetails />} />
               <Route path="/Blog" exact element={<Blog />} />
               <Route path="/auth" element={<Auth />} />
             </Routes>
-          
           </div>
           <Footer />
         </Router>
-      
       </TripFormProvider>
-      </div>
+    </div>
   );
 }
