+++ Frontend/src/Components/TipsDetails/Budget.jsx
