+++ Frontend/src/Components/TipsDetails/PlacesToVisit.jsx
