+++ Frontend/src/Components/TipsDetails/TripDetails.css
