+++ Frontend/src/Components/TipsDetails/Trips.css
